# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "EkMaap onion grading API",
  "description": "SIH26031 - photo-based onion grading. Prototype: see docs/ for limits.",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/dataset/photos": {
   "post": {
    "tags": [
     "dataset"
    ],
    "summary": "Upload Dataset Photo",
    "description": "Add a photo for training/evaluation (not linked to a lot). split_group = e.g. capture day.",
    "operationId": "upload_dataset_photo_api_v1_dataset_photos_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_dataset_photo_api_v1_dataset_photos_post"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "dataset"
    ],
    "summary": "List Dataset",
    "operationId": "list_dataset_api_v1_dataset_photos_get",
    "parameters": [
     {
      "name": "split",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Split"
      }
     },
     {
      "name": "labelled",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Labelled"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/photos/{photo_id}/annotations": {
   "get": {
    "tags": [
     "dataset"
    ],
    "summary": "Get Annotations",
    "operationId": "get_annotations_api_v1_photos__photo_id__annotations_get",
    "parameters": [
     {
      "name": "photo_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Photo Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "dataset"
    ],
    "summary": "Put Annotations",
    "description": "Replace ALL ground-truth labels of a photo (polygons in original photo pixels).",
    "operationId": "put_annotations_api_v1_photos__photo_id__annotations_put",
    "parameters": [
     {
      "name": "photo_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Photo Id"
      }
     },
     {
      "name": "source",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "manual",
       "title": "Source"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "array",
        "items": {
         "$ref": "#/components/schemas/AnnotationIn"
        },
        "title": "Items"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/photos/{photo_id}/prelabel": {
   "post": {
    "tags": [
     "dataset"
    ],
    "summary": "Prelabel",
    "description": "Suggest outlines + conditions with the active models (faster than drawing). save=true stores them as\nannotations (source=prelabel) - a labeller must still check them.",
    "operationId": "prelabel_api_v1_photos__photo_id__prelabel_post",
    "parameters": [
     {
      "name": "photo_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Photo Id"
      }
     },
     {
      "name": "save",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Save"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/dataset/split": {
   "post": {
    "tags": [
     "dataset"
    ],
    "summary": "Assign Split",
    "description": "Assign train/val/test by photo (or by split_group). Stable: a photo keeps its split when more are added.",
    "operationId": "assign_split_api_v1_dataset_split_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SplitIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/dataset/export": {
   "get": {
    "tags": [
     "dataset"
    ],
    "summary": "Export Dataset",
    "description": "ZIP with images/, annotations.json (COCO, categories = conditions, per-image split) and ruler.csv -\nexactly what the scripts/train_*.py and scripts/evaluate.py expect.",
    "operationId": "export_dataset_api_v1_dataset_export_get",
    "parameters": [
     {
      "name": "split",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Split"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/dataset/import": {
   "post": {
    "tags": [
     "dataset"
    ],
    "summary": "Import Coco",
    "description": "Import corrected labels from CVAT / Label Studio (COCO). Images are matched by file name to the\nuploaded photo's original name (or to '<photo id>.<ext>' from our own export). Replaces their labels.",
    "operationId": "import_coco_api_v1_dataset_import_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_coco_api_v1_dataset_import_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/models": {
   "get": {
    "tags": [
     "models"
    ],
    "summary": "List Models",
    "operationId": "list_models_api_v1_models_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "models"
    ],
    "summary": "Register Model",
    "description": "Register a trained model file (from scripts/train_*.py). It is test-loaded before it is accepted.",
    "operationId": "register_model_api_v1_models_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_register_model_api_v1_models_post"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/models/{model_id}/activate": {
   "post": {
    "tags": [
     "models"
    ],
    "summary": "Activate Model",
    "operationId": "activate_model_api_v1_models__model_id__activate_post",
    "parameters": [
     {
      "name": "model_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Model Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/evaluations": {
   "post": {
    "tags": [
     "models"
    ],
    "summary": "Run Evaluation",
    "description": "Evaluate the ACTIVE models + rule set on labelled photos of a split (synchronous; keep test sets modest).",
    "operationId": "run_evaluation_api_v1_evaluations_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EvalIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "models"
    ],
    "summary": "List Evaluations",
    "operationId": "list_evaluations_api_v1_evaluations_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "AnnotationIn": {
    "properties": {
     "polygon": {
      "items": {
       "items": {
        "type": "number"
       },
       "type": "array"
      },
      "type": "array",
      "minItems": 3,
      "title": "Polygon"
     },
     "condition": {
      "type": "string",
      "pattern": "^(sound|sprouted|damaged|rotten)$",
      "title": "Condition"
     },
     "ruler_mm": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMaximum": 200.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Ruler Mm"
     }
    },
    "type": "object",
    "required": [
     "polygon",
     "condition"
    ],
    "title": "AnnotationIn"
   },
...
   "Body_import_coco_api_v1_dataset_import_post": {
    "properties": {
     "coco_file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "Coco File"
     }
    },
    "type": "object",
    "required": [
     "coco_file"
    ],
    "title": "Body_import_coco_api_v1_dataset_import_post"
   },
   "Body_register_model_api_v1_models_post": {
    "properties": {
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "backend": {
      "type": "string",
      "title": "Backend"
     },
     "version": {
      "type": "string",
      "title": "Version"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "metrics": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Metrics",
      "description": "JSON, e.g. the output of scripts/evaluate.py"
     },
     "file": {
      "anyOf": [
       {
        "type": "string",
        "contentMediaType": "application/octet-stream"
       },
       {
        "type": "null"
       }
      ],
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "kind",
     "backend",
     "version"
    ],
    "title": "Body_register_model_api_v1_models_post"
   },
   "Body_upload_dataset_photo_api_v1_dataset_photos_post": {
    "properties": {
     "photo": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "Photo"
     },
     "split_group": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Split Group"
     }
    },
    "type": "object",
    "required": [
     "photo"
    ],
    "title": "Body_upload_dataset_photo_api_v1_dataset_photos_post"
   },
...
   "EvalIn": {
    "properties": {
     "split": {
      "type": "string",
      "pattern": "^(train|val|test)$",
      "title": "Split",
      "default": "test"
     }
    },
    "type": "object",
    "title": "EvalIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "SplitIn": {
    "properties": {
     "train": {
      "type": "number",
      "title": "Train",
      "default": 0.7
     },
     "val": {
      "type": "number",
      "title": "Val",
      "default": 0.15
     },
     "reassign": {
      "type": "boolean",
      "title": "Reassign",
      "default": false
     }
    },
    "type": "object",
    "title": "SplitIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}